Allow logins to existing accounts only with associated accounts
To prevent people can log into existing account with a newly created
social account with a forged email address. We only allow logins with
previously approved service providers.

When a user logs in for the first time, eg. the email does not exists,
then the user is created and the social account is approved
automatically.