Relogin if session is lost or auth_sec_timeout
If the session is lost, possibly to bad server configuration, try a
silent relogin, if an cookie is present that indicates an oauth-session.

If auth_security_timeout is triggered try to re-login based on the
existing session data.

In both cases: Mostly correctly re-set the prvious state after re-login.
Some actions do not work as expected, e.g. pagination in old revisions.

It was decided not to use refresh-tokens. The desired functionality can be achieved by an online-relogin just as
good. Hence this should be prefered to the more user-security invasive
refresh-token mechanism.