Do not attempt another login if the first failed
The user was constantly forwarded to their oauth provider if the login
did not succeed. To fix this the check for a valid session was moved to
the front to handle all non-get requests and requests which are not
directed at doku.php.

Also delete the cookie if the (re-)login fails.