readd support for farmer setups
In a farm, the return from the IDP will land in the farmer and needs to
be redirected to the proper animal. To do so the animal is stored in the
state parameter which all IDPs should pass back to us as is.

Previously this was hardcoded in the Google service only, now it should
work for all services.

Untested though. Also previously the behavior was also loading the
Farmer plugin even when disabled. I don't know why. I changed it.