drop concept of GUID and introduce storageID
Previously we created a new GUID for any user trying to login via oauth.
That GUID was used to store the token states. This polluted the cache
area but even worse, it made it impossible to stay logged in on multiple
devices when the IdP invalidated onld access tokens when issuing a new
one (like google does).

Now a session storage is used for the initial login. Permanent storage
is only initialized after a successful login (based on the user's email
address).