do not clear login data on every load
logoff is called a lot when a user is not authenticated yet. This can
interfere with the oauth login process when calls are made while the
user is at the IdP, for example triggered by AJAX beacons like created
by the statistics plugin.
There is no real reason to delete the temporary session data. We can
simply keep it until it is used or replaced by new data.